refactor: reorganize TodoTab component structure and update import paths
- Moved TodoTab component to a new directory structure under `components/todo/components/`.
- Updated import paths in `MaxExpandContentLazy.tsx` to reflect the new location of TodoTab.
- Deleted the old TodoTab file and created a new one in the updated directory, maintaining the original functionality and code structure.
- Added a new index file for the Todo component to facilitate easier imports.
