+++ src/renderer/components/components/StandaloneWindowViewport.tsx
@@ -25,7 +25,7 @@
  */
 
 import type { JSX } from 'react';
-import { TodoTab } from '../states/maxExpand/components/TodoTab';
+import { TodoTab } from '../states/maxExpand/components/todo/components/TodoTab';
 import { CountdownTab } from '../states/maxExpand/components/countdown';
 import { UrlFavoritesTab } from '../states/maxExpand/components/UrlFavoritesTab';
 import { AlbumTab } from '../states/maxExpand/components/album/components/AlbumTab';
