+++ src/renderer/components/states/maxExpand/MaxExpandContentEager.tsx
@@ -28,7 +28,7 @@ import type { ReactElement } from 'react';
 import type { MaxExpandTab } from '../../../store/types';
 import { MaxExpandContentShell } from './MaxExpandContentShell';
 import { AiChatTab } from './components/agent/components/AiChatTab';
-import { TodoTab } from './components/TodoTab';
+import { TodoTab } from './components/todo/components/TodoTab';
 import { UrlFavoritesTab } from './components/UrlFavoritesTab';
 import { LocalFileSearchTab } from './components/LocalFileSearchTab';
 import { ClipboardHistoryTab } from './components/clipBoardHistory';
